refactor(article,series): integrate series application to article application
