+++ services/article/application/src/main/java/nettee/blolet/article/application/port/SeriesArticleCommandRepositoryPort.java
@@ -1,4 +1,4 @@
-package nettee.series.article.application.port;
+package nettee.blolet.article.application.port;
 
 import nettee.blolet.article.domain.SeriesArticle;
