+++ services/article/application/src/main/java/nettee/blolet/article/application/port/SeriesQueryRepositoryPort.java
@@ -1,4 +1,4 @@
-package nettee.series.application.port;
+package nettee.blolet.article.application.port;
 
 import nettee.blolet.article.readmodel.SeriesQueryModels.SeriesDetail;
 import nettee.blolet.article.readmodel.SeriesQueryModels.SeriesSummary;
