+++ services/article/application/src/main/java/nettee/blolet/article/application/service/SeriesArticleQueryService.java
@@ -1,9 +1,9 @@
-package nettee.series.article.application.service;
+package nettee.blolet.article.application.service;
 
 import lombok.RequiredArgsConstructor;
+import nettee.blolet.article.application.port.SeriesArticleQueryRepositoryPort;
+import nettee.blolet.article.application.usecase.SeriesArticleReadUseCase;
 import nettee.blolet.article.readmodel.SeriesArticleQueryModels.SeriesArticleSummary;
-import nettee.series.article.application.port.SeriesArticleQueryRepositoryPort;
-import nettee.series.article.application.usecase.SeriesArticleReadUseCase;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
