+++ services/article/application/src/main/java/nettee/blolet/article/application/usecase/SeriesDeleteUseCase.java
@@ -1,4 +1,4 @@
-package nettee.series.application.usecase;
+package nettee.blolet.article.application.usecase;
 
 public interface SeriesDeleteUseCase {
