+++ services/article/application/src/main/java/nettee/blolet/article/application/usecase/SeriesReadUseCase.java
@@ -1,4 +1,4 @@
-package nettee.series.application.usecase;
+package nettee.blolet.article.application.usecase;
 
 import nettee.blolet.article.readmodel.SeriesQueryModels.SeriesDetail;
 import nettee.blolet.article.readmodel.SeriesQueryModels.SeriesSummary;
