+++ services/series/driving/web-mvc/src/main/java/nettee/series/driving/web/SeriesQueryApi.java
@@ -6,11 +6,11 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
+import nettee.blolet.article.application.usecase.SeriesReadUseCase;
+import nettee.blolet.article.application.usecase.SeriesVisitUseCase;
 import nettee.blolet.article.readmodel.SeriesQueryModels.SeriesDetail;
 import nettee.blolet.jwt.filter.annotation.AuthUser;
 import nettee.blolet.jwt.filter.annotation.AuthorizedUser;
-import nettee.series.application.usecase.SeriesReadUseCase;
-import nettee.series.application.usecase.SeriesVisitUseCase;
 import nettee.series.driving.web.dto.SeriesQueryDto.SeriesDetailResponse;
 import nettee.series.driving.web.dto.SeriesQueryDto.SeriesSummaryResponse;
 import org.springframework.web.bind.annotation.GetMapping;
